python-adk-mcp/pyproject.toml
[project]
name = "graph_agent"
version = "0.1.0"
description = "Knowledge Graph agent with ADK"
authors = [{ name = "Niko Raes", email = "niko.raes@konnektr.io" }]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "google-adk>=1.8.0",
    "langchain>=0.3.0,<1.2.0",
    "google-cloud-aiplatform[agent-engines,evaluation]>=1.93.0",
    "langchain-community>=0.3.25",
    "python-dotenv>=1.1.0",
    "stackapi>=0.3.1",
    "toolbox-core>=0.1.0",
]

[dependency-groups]
dev = ["pytest>=8.3.5", "pytest-asyncio>=0.26.0", "agent-starter-pack>=0.14.1"]

[project.optional-dependencies]
lint = [
    "ruff>=0.4.6",
    "mypy>=1.15.0",
    "codespell>=2.2.0",
    "types-pyyaml>=6.0.12.20240917",
    "types-requests>=2.32.0.20240914",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901"] # ignore line too long, too complex

[tool.ruff.lint.isort]
known-first-party = ["software_bug_assistant"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]

exclude = [".venv"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""

# This configuration file is used by goo.gle/agent-starter-pack to power remote templating.
# It defines the template's properties and settings.
[tool.agent-starter-pack]
example_question = "Can you create a knowledge graph of this text please?"

[tool.agent-starter-pack.settings]
agent_directory = "graph_agent"
